% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qss-student.R
\name{list_psets}
\alias{list_psets}
\title{List available problem sets}
\usage{
list_psets()
}
\value{
a vector of names of bundled problem sets
}
\description{
Call this function to get the names of all problem sets
in the package.
}
