% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qss-student.R
\name{preview_pset}
\alias{preview_pset}
\title{Preview a problem set}
\usage{
preview_pset(pname)
}
\arguments{
\item{pname}{Name of a problem set}
}
\value{
Nothing.
}
\description{
Launches the pdf viewer to show the questions in a problem set. If the
problem set name is not recognized a menu of options is provided.
}
