% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qss-student.R
\docType{package}
\name{qss-student}
\alias{qss-student}
\alias{qss-student-package}
\title{fsi: POL 245 problem sets at FSI}
\description{
This package contains problem sets from the qss-inst repository
of exercises in Imai 2016 An Introduction to Quantitative Social Science.
}
\details{
The instructor assigns students the name of a problem set, e.g.
'gay-marriage'.  Each student loads the package and calls

\code{get_pset('gay-marriage')}

This function unzips the problem set into the student's current working
directory, and switches their current working directory to the top level of
the problem set.

Students open the Rmd file and add their own text and code blocks to answer
the questions.  They then submit their compiled document as homework.

A problem set from the qss-inst repository always contains
\itemize{
 \item{an Rmd file describing the data set and listing several question.}
 \item{a pdf version of the Rmd file for readability}
 \item{a data folder containing the data set}
}
}
