% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qss-student.R
\name{unzip_pset_in_wd}
\alias{unzip_pset_in_wd}
\title{Unzips problem set in working directory}
\usage{
unzip_pset_in_wd(fname, setwd, overwrite)
}
\arguments{
\item{fname}{Problem set name}

\item{setwd}{Whether to set the working directory to the unpacked directory}

\item{overwrite}{Whether to overwrite any existing versions of the same
problem set.}
}
\value{
Nothing
}
\description{
Unzips problem set in working directory
}
