% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi.R
\name{get_pset}
\alias{get_pset}
\title{Locate a problem set by name and unpack it}
\usage{
get_pset(pname, setwd = TRUE)
}
\arguments{
\item{pname}{Name of a problem set}

\item{setwd}{Whether to change working directory to where the files were unpacked. Default: TRUE}
}
\value{
None
}
\description{
This function \itemize{
  \item{Finds a problem set by name},
  \item{Unpacks it to wherever you are working},
  \item{Moves your current working directory to the problem set folder (unless you set \code{setwd} to FALSE).}
  \item{Reminds you that you probably want to make the Files view match your new working directory, and tells you how.}
}
}
\examples{
\dontrun{ get_pset("test") }

}
